Modularize shared note snapshot routes
Separate shared-note snapshot transport, mutation validation, and OpenAPI wiring from CloudSync credential routes without changing API behavior.
